Minimal horrible hacks to investigate some normalization work
The router needs to perform normalization on queries. It seems like we
should be able to leverage some support for this from the parser. These
horrible hacks represent some of the kind of work that would be needed.

This is *not* intended to proceed beyond being a discussion point for
real changes...